+++ src/main/java/com/atlauncher/gui/tabs/settings/GeneralSettingsTab.java
@@ -29,6 +29,7 @@ import javax.swing.JPanel;
 
 import com.atlauncher.App;
 import com.atlauncher.FileSystem;
+import com.atlauncher.builders.HTMLBuilder;
 import com.atlauncher.data.Constants;
 import com.atlauncher.data.Language;
 import com.atlauncher.gui.components.JLabelWithHover;
@@ -62,6 +63,8 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
     private JCheckBox enableFeralGamemode;
     private JLabelWithHover enablePackTagsLabel;
     private JCheckBox enablePackTags;
+    private JLabelWithHover disableAddModRestrictionsLabel;
+    private JCheckBox disableAddModRestrictions;
 
     public GeneralSettingsTab() {
         // Language
@@ -241,6 +244,7 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
         add(enableDiscordIntegration, gbc);
 
         // Enable Feral Gamemode
+      
         if (OS.isLinux()) {
             boolean gameModeExistsInPath = Utils.executableInPath("gamemoderun");
 
@@ -272,7 +276,7 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
             add(enableFeralGamemode, gbc);
         }
 
-        // Pack tags
+        // Enable Pack Tags
 
         gbc.gridx = 0;
         gbc.gridy++;
@@ -289,6 +293,25 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
         enablePackTags = new JCheckBox();
         enablePackTags.setSelected(App.settings.enabledPackTags());
         add(enablePackTags, gbc);
+
+        // Disable Curse Minecraft version restrictions
+
+        gbc.gridx = 0;
+        gbc.gridy++;
+        gbc.insets = LABEL_INSETS;
+        gbc.anchor = GridBagConstraints.BASELINE_TRAILING;
+        disableAddModRestrictionsLabel = new JLabelWithHover(GetText.tr("Disable Add Mod Restrictions"), HELP_ICON,
+                new HTMLBuilder().center().text(GetText.tr(
+                        "This will allow you to disable the restrictions in place to prevent you from installing mods from Curse that are not for your current Minecraft version or loader.<br/><br/>By disabling these restrictions, you can install any mod, so be sure that it's compatable with the Minecraft version and loader (if any) that you're on."))
+                        .build());
+        add(disableAddModRestrictionsLabel, gbc);
+
+        gbc.gridx++;
+        gbc.insets = FIELD_INSETS;
+        gbc.anchor = GridBagConstraints.BASELINE_LEADING;
+        disableAddModRestrictions = new JCheckBox();
+        disableAddModRestrictions.setSelected(App.settings.disabledAddModRestrictions());
+        add(disableAddModRestrictions, gbc);
     }
 
     public boolean needToReloadTheme() {
@@ -320,6 +343,7 @@ public class GeneralSettingsTab extends AbstractSettingsTab {
         }
 
         App.settings.setPackTags(enablePackTags.isSelected());
+        App.settings.setAddModRestrictions(disableAddModRestrictions.isSelected());
     }
 
     @Override