+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Add in some more error checking for project id's from slugs
 - Fix not being able to get some Curse modpack slugs
+- Fix 1.13+ Forge packs on Curse not installing as 1.13+ packs