+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Add in ability to download worlds from CurseForge
 - Fix mods from curse able to be installed with no loader
+- Add setting to disable add mod restrictions for instances