+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - Add in ability to download worlds from CurseForge
 - Fix mods from curse able to be installed with no loader
 - Add setting to disable add mod restrictions for instances
+- Small tweak to save password warning (https://github.com/ATLauncher/ATLauncher/pull/376)
+- Add Feral gamemode integration (https://github.com/ATLauncher/ATLauncher/pull/377/)