+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.3.3.0
 
 - Add right click menu on Edit Mods dialog
+- Fix issue with quotes in path causing runtime issues